Revert "🎁 Interact by location/place - Issue #131 - v1.3.0"
